First iteration of using oauth client keys to mint an auth key
Since this is supposed to always bring up a tailscale node, we can't
rely on people rotating auth keys regularly. So to facilitate that,
let's use tailscale's oauth keys and mint a fresh preauth access
token.

As a bonus, we can detect if there is a node that conflicts & remove
it? Maybe this is a thing.