Handle credentials in a cleaner way
Extract the credential retrieval into a function, getCredential that
will retrieve the cred from the environment & remove the variable on
successful retrieval (this should result in a cleaner environment in
downstream processes).

Also, support use of the $CREDENTIALS_DIRECTORY that systemd supports
for its LoadCredential directive; this allows handling them in a far
more secure way (not exposing them on the env, giving only the service
user access).