Correct error handling on exit when serving prometheus metrics
The logic for serving prometheus metrics was somewhat broken: It would
attempt to check that the cancellation context didn't indicate that we
should exit, but it would only do that _before_ starting to serve.

Instead, we must serve, and if serve returns a fault, check whether
we're meant to exit. Oops.