Improve reliability of nix tests
Previously, we'd import an offline-generated headscale DB dump containing api and preauth keys that hoopsnake can use. That's sadly not viable now that headscale is upgraded in nixpkgs-unstable.

Instead, let's generate the keys in an "offline" headscale at build time and take a db dump there. It's just as gross, but at least it'll not be breaking as easily!