test: Remove logic that pings alice-boot
We're doing exactly the same thing in the line below, but without a timeout (and the timeout gets hit in CI often, due to slowness of the runners).


Add a -configtest commandline flag

It's scary rebooting into an initrd where hoopsnake might not be able to come up / mint an authkey. This lays the foundation for checking if HS can work on configuration rollouts.