Detach tsnet cancellation from SSH server termination
Shutting down the tsnet interface at the same time as the SSH server results in connected clients not getting the notification that their connection was terminated.

So, create a child context for the tsnet service that doesn't cancel immediately; instead, set it up with its own cancellation handler that gets invoked once the SSH server is shut down.